pyproject.toml
[tool.poetry]
name = "pr-reviewer-service"
version = "1.0.0"
packages = [
    { include = "app" }
]
description = "Сервис назначения ревьюеров для Pull Request'ов"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
python-multipart = "^0.0.12"
pyyaml = "^6.0.2"
sqlalchemy = "^2.0.35"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
redis = "^5.2.0"
python-dotenv = "^1.0.1"
greenlet = "^3.2.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"
black = "^24.8.0"
ruff = "^0.6.0"
faker = "^24.0.0"
pytest-mock = "^3.14.0"
aiosqlite = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

